A compose and makefile
This uses a Docker image from the community to manage Jekyll sites.

If you are checking this out for the first time, do a

`make new`

in the repository. This will add a bunch of pieces to turn it into a
Jekyll site (and therefore a GH Pages site).

To preview the site locally, do a

`make preview`

which is the same as saying

`docker compose up`

and, things should magically happen.